% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actab.R
\name{actab}
\alias{actab}
\title{Return a \emph{gt} table displaying contingency analysis results (from two categorical variables of a dataframe)}
\usage{
actab(
  data,
  var1,
  var2,
  weight,
  option,
  palette,
  title,
  subtitle,
  var1.lib,
  var2.lib,
  source,
  lecture
)
}
\arguments{
\item{data}{a dataframe with \emph{n} rows (individuals) and \emph{p} columns (including categorical variables). If \emph{weight} is given, this dataframe must also contain a numeric variable indicating row weights. To avoid misunderstandings, categorical variables must not contain any NA value}

\item{var1}{colname of the independent variable}

\item{var2}{colname of the dependent variable}

\item{weight}{an optional numeric column of \emph{data} indicating row weights (if weight is informed, only option=0 is allowed)}

\item{option}{indicates which table to display : \emph{0} (row percentages), \emph{1} (residuals and Chi2 contributions) or \emph{2} (row percentages including some contingency analysis results). Only 0 is allowed for weighted data.}

\item{palette}{indicates which palette to use for "option=1" : "GreenRed" or "RedGreen"}

\item{title}{title to display}

\item{subtitle}{subtitle to display}

\item{var1.lib}{\emph{var1}'s name to display}

\item{var2.lib}{\emph{var2}'s name to display}

\item{source}{\emph{source} to display}

\item{lecture}{only available if option=0 : an example of how to read values in the table}
}
\value{
a \emph{gt} table : residuals and Chi2 contributions if \emph{option==1} ; row percentages including some contingency analysis results if \emph{option==2}
}
\description{
For more details : the Github repository at \url{https://github.com/alietteroux/exploquantir}
}
