% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acXvars.R
\name{acXvars}
\alias{acXvars}
\title{Calculate and plot Cramer's V between several categorical variables}
\usage{
acXvars(data, weight, excl, progress, graph, plot.p.value)
}
\arguments{
\item{data}{a dataframe with \emph{n} rows (individuals) and \emph{p} columns (categorical variables). If \emph{weight} is given, this dataframe must also contain a numeric variable indicating row weights. To avoid misunderstandings, categorical variables must not contain any NA value}

\item{weight}{an optional numeric column of \emph{data} indicating row weights}

\item{excl}{an optional list or dataframe indicating wich levels of some categorical variables shoul be ignored. If it's a dataframe : it must contained a column named "variable" and a column named "modalite". If it's a list : this list must be nested with 2 levels (level 1 = variable, level 2 = modalite). See examples.}

\item{progress}{boolean, if TRUE displays progress on the console}

\item{graph}{boolean, if TRUE a graph is displayed}

\item{plot.p.value}{boolean, if TRUE a p-value threshold (0.05) is displayed on the graph}
}
\value{
a list with two components :
\itemize{
\item \strong{d.ac} : a dataframe that contains as many rows as contingency analysis have been realised.
\itemize{
\item \emph{var1} and \emph{var2} (categorical variables)
\item \emph{sum.tc} : sum of the contingency table. Nota Bene : this sum is not necessarily equals to the number of \emph{data}'s individuals when levels are excluded by \emph{excl}
\item \emph{var1_mod_m5} and \emph{var2_mod_m5} : levels of \emph{var1} and \emph{var2} matching less than 5 individuals
\item \emph{part_cases_sup5} : percent of cells of the contingency table matching less than 5 individuals
\item \emph{v.cramer} : Cramer's V calculated by \emph{questionr::chisq.test()} when \emph{weight} is not given (Pearson's Chi-squared test) ; by \emph{survey::svychisq()} when \emph{weight} is given (Pearson's X^2: Rao & Scott adjustment)
\item \emph{v.cramer.cl} : category of Cramer's : >= 0.3 ; >= 0.15 et <0.3 ; >= 0.05 et <0.15 ; < 0.5
\item \emph{cond.Grasland} : indicates whether Claude Grasland's conditions are met (O) or not (N), that means : \emph{sum.tc}>20 ; \emph{var1_mod_m5} and \emph{var2_mod_m5} are null ; \emph{part_cases_sup5}>=80
\item \emph{croisement.ordre.alphab} : allows to differenciate each crossing
\item \emph{p.value} : p.value of Chi-squared test
}
\item \strong{plot} : a \emph{ggplot} graphic to quickly visualize strongest statistic relationships
}
}
\description{
For more details : the Github repository at \url{https://github.com/alietteroux/exploquantir}
}
