% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aovXvars.R
\name{aovXvars}
\alias{aovXvars}
\title{Calculate and plot proportions of numeric variables variances explained by categorical variables}
\usage{
aovXvars(data, vars.quali, vars.quanti, progress, graph)
}
\arguments{
\item{data}{a dataframe with \emph{n} rows (individuals) and \emph{p} columns (categorical and numeric variables). To avoid misunderstandings, categorical variables must not contain any NA value}

\item{vars.quali}{names of categorical variables of \emph{data}}

\item{vars.quanti}{names of numeric variables of \emph{data}}

\item{progress}{boolean, if TRUE displays progress on the console}

\item{graph}{indicates which graph to display : \emph{intens} (proportions of variances explained by categorical variables), \emph{detail} (boxplots for each categorical x numeric variables crossing), \emph{none} (no plot is displayed)}
}
\value{
a list with two components :
\itemize{
\item \strong{d.aov} : a dataframe that contains as many rows as analysis of variance (ANOVA) have been realised.
\itemize{
\item \emph{var.quali} : categorical variable of the ANOVA
\item \emph{var.quanti} : numeric variable of the ANOVA
\item \emph{sd.eff.gr} : standard deviation of individuals'numbers in each group of the categorical variable
\item \emph{cv.eff.gr} : coefficient of variation of individuals'numbers in each group of the categorical variable
\item \emph{V1} : sum square of intergroup variance
\item \emph{V2} : sum square of intragroup variance
\item \emph{intensite} : percent of \emph{var.quanti} variance explained by \emph{var.quali}
\item \emph{p.value} : p.value of Fisher's test
}
\item \strong{plot.intens} : a \emph{ggplot} graphic to quickly visualize strongest statistic relationships
\item \strong{plot.detail} : a \emph{ggplot} graphic that displays a boxplot for each \emph{categorical x numeric} variables crossing
}
}
\description{
For more details : the Github repository at \url{https://github.com/alietteroux/exploquantir}
}
