% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploquantir-package.R
\docType{package}
\name{exploquantir-package}
\alias{exploquantir}
\alias{exploquantir-package}
\title{exploquantir: Functions to make easier exploratory analysis of quantitative data in social sciences}
\description{
This package goes with PROGEDO-Loire formations (https://msh-progedo-loire.univ-nantes.fr/). It aims to make easier exploratory analysis of quantitative data in social sciences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alietteroux/exploquantir}
  \item Report bugs at \url{https://github.com/alietteroux/exploquantir/issues}
}

}
\author{
\strong{Maintainer}: Aliette Roux \email{aliette.roux@univ-nantes.fr}

}
\keyword{internal}
