% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabToGetIndexACM.R
\name{TabToGetIndexACM}
\alias{TabToGetIndexACM}
\title{Return a helpful \emph{dataframe} for specific MCA with FactoMineR}
\usage{
TabToGetIndexACM(data.vars, vars.quali.supp)
}
\arguments{
\item{data.vars}{a dataframe with \emph{n} rows (individuals) and \emph{p} columns of categorical variables, from which to achieve a MCA.}

\item{vars.quali.supp}{a vector listing colnames of the categorical supplementary variables in \emph{data.vars}.}
}
\value{
a \emph{dataframe} from which we can easily get indexes of "junk" categories, so as to achieve a specific MCA with FactoMineR.
}
\description{
For more details : the Github repository at \url{https://github.com/alietteroux/exploquantir}
}
