% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabToPlotACM.R
\name{TabToPlotACM}
\alias{TabToPlotACM}
\title{Return a \emph{dataframe} that makes easier MCA-plotting with ggplot}
\usage{
TabToPlotACM(data.ACM, res.ACM, vars.quali.supp, code.ex)
}
\arguments{
\item{data.ACM}{a dataframe with \emph{n} rows (individuals) and \emph{p} columns of categorical variables.}

\item{res.ACM}{a "MCA" object : the result of \emph{FactoMineR:MCA(data.ACM)}.}

\item{vars.quali.supp}{a vector listing colnames of the categorical supplementary variables in \emph{data.ACM}.}

\item{code.ex}{boolean, if TRUE displays an example of ggplot code in the Console.}
}
\value{
\itemize{
\item a \emph{dataframe} from which we can easily plot a MCA graph (it must be named "modalites" to work with \emph{code.ex})
\item an example of ggplot code in the Console (if code.ex=T),
}
}
\description{
This function is inspired by \href{https://quanti.hypotheses.org/930}{Anton Perdoncin, "\emph{Représenter graphiquement les résultats d’une analyse factorielle avec R}"}.
}
\details{
For more details : the Github repository at \url{https://github.com/alietteroux/exploquantir}
}
